Fix warnings about allocations exceeding maximum object size
In member function 'Init',
    inlined from 'Init' at .../src/Xrd/XrdBuffXL.cc:64:6:
.../src/Xrd/XrdBuffXL.cc:90:30: warning: argument 1 value '18446744073709551615' exceeds maximum object size 9223372036854775807 [-Walloc-size-larger-than=]
   90 |    bucket = new BuckVec[lg2+1];
      |                              ^
/usr/include/c++/12/new: In member function 'Init':
/usr/include/c++/12/new:128:26: note: in a call to allocation function 'operator new []' declared here
  128 | _GLIBCXX_NODISCARD void* operator new[](std::size_t) _GLIBCXX_THROW (std::bad_alloc)
      |                          ^

.../src/XrdOuc/XrdOucBuffer.cc: In member function '__ct_base ':
.../src/XrdOuc/XrdOucBuffer.cc:79:30: warning: argument 1 value '18446744073709551615' exceeds maximum object size 9223372036854775807 [-Walloc-size-larger-than=]
   79 |    bSlot = new BuffSlot[slots];
      |                              ^
/usr/include/c++/12/new:128:26: note: in a call to allocation function 'operator new []' declared here
  128 | _GLIBCXX_NODISCARD void* operator new[](std::size_t) _GLIBCXX_THROW (std::bad_alloc)
      |                          ^

.../src/XrdNet/XrdNetUtils.cc: In function 'GetAddrs':
.../src/XrdNet/XrdNetUtils.cc:262:35: warning: argument 1 value '18446744073709551615' exceeds maximum object size 9223372036854775807 [-Walloc-size-larger-than=]
  262 |        *aVec = new XrdNetAddr[aVsz];
      |                                   ^
/usr/include/c++/12/new:128:26: note: in a call to allocation function 'operator new []' declared here
  128 | _GLIBCXX_NODISCARD void* operator new[](std::size_t) _GLIBCXX_THROW (std::bad_alloc)
      |                          ^

In member function '__ct ',
    inlined from '__ct_base ' at .../src/XrdXrootd/XrdXrootdJob.cc:459:29:
.../src/./XrdOuc/XrdOucTable.hh:43:30: warning: argument 1 value '18446744073709551615' exceeds maximum object size 9223372036854775807 [-Walloc-size-larger-than=]
   43 |                      Table = new OucTable[maxe];
      |                              ^
/usr/include/c++/12/new: In member function '__ct_base ':
/usr/include/c++/12/new:128:26: note: in a call to allocation function 'operator new []' declared here
  128 | _GLIBCXX_NODISCARD void* operator new[](std::size_t) _GLIBCXX_THROW (std::bad_alloc)
      |                          ^

.../src/XrdPosix/XrdPosixAdmin.cc: In member function 'FanOut':
.../src/XrdPosix/XrdPosixAdmin.cc:69:27: warning: argument 1 value '18446744073709551615' exceeds maximum object size 9223372036854775807 [-Walloc-size-larger-than=]
   69 |    uVec = new XrdCl::URL[i];
      |                           ^
/usr/include/c++/12/new:128:26: note: in a call to allocation function 'operator new []' declared here
  128 | _GLIBCXX_NODISCARD void* operator new[](std::size_t) _GLIBCXX_THROW (std::bad_alloc)
      |                          ^

+++ src/XrdNet/XrdNetUtils.cc
@@ -259,7 +259,7 @@ const char  *XrdNetUtils::GetAddrs(const char            *hSpec,
 //
    if (aInfo.aNum4 || aInfo.aNum6)
       {aVsz = aInfo.aNum4 + aInfo.aNum6;
-       *aVec = new XrdNetAddr[aVsz];
+       *aVec = new XrdNetAddr[(unsigned int)aVsz];
        FillAddr(aInfo, *aVec);
       }
